Expose cue-parsing utilities
